Minor rename: sentence_ranges -> annotation (#134)
