+++ src/translator/annotation.cpp
@@ -1,4 +1,4 @@
-#include "sentence_ranges.h"
+#include "annotation.h"
 #include <cassert>
 #include <iostream>
