+++ src/translator/annotation.h
